Move scripts to mailing_lists directory     -update imports/etc.
