+++ website/settings/defaults.py
@@ -397,7 +397,7 @@ else:
             'kwargs': {'dry_run': False}
         },
         'update-project-mailing-lists': {
-            'task': 'scripts.update_project_mailing_lists',
+            'task': 'scripts.mailing_lists.update_project_mailing_lists',
             'schedule': crontab(minute=0, hour=3, day_of_week=0),  # Sunday 3:00 a.m.
             'kwargs': {'dry_run': False},
         },
